requirements.txt
torch
torchvision
gradio
pillow
numpy
opencv-python
matplotlib
seaborn